Edited tests, scattered them in different files
