+++ CoCowork.BusinessLayer/Services/IService/IMiniOfficeService.cs
