--- CoCowork.Tests/ItemOrderServiceTests/LaptopOrderServiveTests.cs
@@ -1,6 +1,0 @@
-﻿namespace CoCowork.BusinessLayer.Tests
-{
-    internal class LaptopOrderServiveTests
-    {
-    }
-}
