+++ CoCowork.Tests/ItemOrderServiceTests/OrderServiceTests.cs
@@ -10,27 +10,13 @@ namespace CoCowork.BusinessLayer.Tests
 {
     public class OrderServiceTests
     {
-        private readonly Mock<ProductRepository> _productRepositoryMock;
-
-        private readonly Mock<IMiniOfficeOrderRepository> _miniOfficeRepositoryMock;
         private readonly Mock<IOrderRepository> _orderRepositoryMock;
-        private readonly Mock<IPlaceOrderRepository> _placeOrderRepository;
-        private readonly Mock<IProductOrderRepository> _productOrderRepository;
-        private readonly Mock<ILaptopOrderRepository> _laptopOrderRepositoryMock;
-        private readonly Mock<IRoomOrderRepository> _roomOrderRepository;
-        
-
         private readonly OrdersTestData _ordersTestData;
         private readonly ClientTestData _clientTestData;
 
         public OrderServiceTests()
         {
             _orderRepositoryMock = new Mock<IOrderRepository>();
-            _laptopOrderRepositoryMock = new Mock<ILaptopOrderRepository>();
-            _miniOfficeRepositoryMock = new Mock<IMiniOfficeOrderRepository>();
-            _placeOrderRepository = new Mock<IPlaceOrderRepository>();
-            _productOrderRepository = new Mock<IProductOrderRepository>();
-            _roomOrderRepository = new Mock<IRoomOrderRepository>();
             _clientTestData = new ClientTestData();
             _ordersTestData = new OrdersTestData();
         }
@@ -51,81 +37,5 @@ namespace CoCowork.BusinessLayer.Tests
             //assert
             _orderRepositoryMock.Verify(m => m.Add(It.IsAny<Order>()), Times.Once());
         }
-
-        [Test]
-        public void AddLaptopOrder()
-        {
-            //arrange
-            var laptop = _ordersTestData.GetLaptopModelForTests();
-            var sut = new LaptopService(_laptopOrderRepositoryMock.Object);
-
-            _laptopOrderRepositoryMock.Setup(m => m.Add(It.IsAny<LaptopOrder>()));
-            ////act
-            var actual = sut.AddItemOrder(laptop);
-
-            //assert
-            _laptopOrderRepositoryMock.Verify(m => m.Add(It.IsAny<LaptopOrder>()), Times.Once());
-        }
-
-        [Test]
-        public void AddProductOrder()
-        {
-            //arrange
-            var product = _ordersTestData.GetProductModelForTests();
-            _productOrderRepository.Setup(m => m.Add(It.IsAny<ProductOrder>()));
-            var sut = new ProductService(_productOrderRepository.Object);
-
-            ////act
-            var actual = sut.AddItemOrder(product);
-
-            //assert
-            _productOrderRepository.Verify(m => m.Add(It.IsAny<ProductOrder>()), Times.Once());
-        }
-
-        [Test]
-        public void AddPlaceOrder()
-        {
-            //arrange
-            var place = _ordersTestData.GetPlaceModelForTests();
-            _placeOrderRepository.Setup(m => m.Add(It.IsAny<PlaceOrder>()));
-            var sut = new PlaceService(_placeOrderRepository.Object);
-
-            ////act
-            var actual = sut.AddItemOrder(place);
-
-            //assert
-            _placeOrderRepository.Verify(m => m.Add(It.IsAny<PlaceOrder>()), Times.Once());
-        }
-
-        [Test]
-        public void AddMiniOfficeOrder()
-        {
-            //arrange
-            var miniOffice = _ordersTestData.GetMiniOfficeModelForTests();
-            _miniOfficeRepositoryMock.Setup(m => m.Add(It.IsAny<MiniOfficeOrder>()));
-            var sut = new MiniOfficeService(_miniOfficeRepositoryMock.Object);
-
-            ////act
-            var actual = sut.AddItemOrder(miniOffice);
-
-            //assert
-            _miniOfficeRepositoryMock.Verify(m => m.Add(It.IsAny<MiniOfficeOrder>()), Times.Once());
-        }
-
-        [Test]
-        public void AddRoomOrder()
-        {
-            //arrange
-            var roomOffice = _ordersTestData.GetRoomModelForTests();
-            _roomOrderRepository.Setup(m => m.Add(It.IsAny<RoomOrder>()));
-            var sut = new RoomService(_roomOrderRepository.Object);
-
-            ////act
-            var actual = sut.AddItemOrder(roomOffice);
-
-            //assert
-            _roomOrderRepository.Verify(m => m.Add(It.IsAny<RoomOrder>()), Times.Once());
-        }
-
     }
 }
